Permit message entities to expose dissector tables
Permit message entities to expose a dissector table to allow generic
messages that encapsulate another protocol within a protocol definition.

This does not support nesting protocols directly, however. To define
multiple protocols, it is still necessary to define them a separate
files (which, in turn, will be generated as separate Wireshark plugins).